From: Dan McGuirk <mcguirk@indirect.com> Subject: [HACKERS] linux/alpha patches
These patches lay the groundwork for a Linux/Alpha port.  The port doesn't
actually work unless you tweak the linker to put all the pointers in the
first 32 bits of the address space, but it's at least a start.  It
implements the test-and-set instruction in Alpha assembly, and also fixes
a lot of pointer-to-integer conversions, which is probably good anyway.

+++ src/backend/bootstrap/bootstrap.c
@@ -7,7 +7,7 @@
  * Copyright (c) 1994, Regents of the University of California
  *
  * IDENTIFICATION
- *    $Header: /cvsroot/pgsql/src/backend/bootstrap/bootstrap.c,v 1.15 1997/01/24 22:42:30 scrappy Exp $
+ *    $Header: /cvsroot/pgsql/src/backend/bootstrap/bootstrap.c,v 1.16 1997/03/12 20:57:33 scrappy Exp $
  *
  *-------------------------------------------------------------------------
  */
@@ -485,7 +485,7 @@ boot_openrel(char *relname)
     
     if (!Quiet)
         printf("Amopen: relation %s. attrsize %d\n", relname,
-               ATTRIBUTE_TUPLE_SIZE);
+               (int)ATTRIBUTE_TUPLE_SIZE);
     
     reldesc = heap_openr(relname);
     Assert(reldesc);
